Refactor(App, ListItem, ProductInput): Cambiar de lugar la instanciacion del objeto
